add types in compare_abs

+++ src/Helpers/helpers.c
@@ -2,6 +2,7 @@
 
 #include "../s21_decimal.h"
 
+
 int s21_get_bit(s21_decimal decimal, int bit_position) {
   int byte_position = bit_position / 32;
   int bit_offset = bit_position % 32;
@@ -202,9 +203,9 @@ int s21_compare_abs(s21_decimal value_1, s21_decimal value_2) {
 
   // Сравнение мантисс (игнорируя знаки)
   for (int i = 2; i >= 0; --i) {
-      if (value_1.bits[i] < value_2.bits[i]) {
+      if ((uint32_t)value_1.bits[i] < (uint32_t)value_2.bits[i]) {
           return -1;  // value_1 < value_2 по модулю
-      } else if (value_1.bits[i] > value_2.bits[i]) {
+      } else if ((uint32_t)value_1.bits[i] > (uint32_t)value_2.bits[i]) {
           return 1;   // value_1 > value_2 по модулю
       }
   }
@@ -233,10 +234,8 @@ int s21_subtract_core(s21_decimal value_1, s21_decimal value_2, s21_decimal *res
   if (!result) return AR_NAN;
 
   s21_zero_decimal(result);
-
-  // Проверяем, какое число больше по модулю
   if (s21_compare_abs(value_1, value_2) < 0) {
-      s21_decimal temp = value_1;
+    s21_decimal temp = value_1;
       value_1 = value_2;
       value_2 = temp;
   }
